Update Misc.scala
bsp location to be user settable. Currently bsp and application folder are at different levels in root folder and very difficult for user to navigate. If it is under one directory under the board it would be convenient. 
Made changes in digilent and radiona boards to take care of this. These boards continue to use the old hierarchy.